// frontend/src/app/modules/landing/admin/models/ResponseModel.ts
export interface ResponseModel {
    current_page: number
    data: any[]
    first_page_url: string
    from: number
    last_page: number
    last_page_url: string
    links: any[]
    next_page_url: string
    path: string
    per_page: number
    prev_page_url: string
    to: number
    total: number
  }
  